preparation methode pour modifier le mot de passe methode pour anonymiser un membre methodes pour supprimer un club/un terrain/un membre

+++ src/main/java/be/company/fca/controller/ChampionnatController.java
@@ -48,11 +48,12 @@ public class ChampionnatController {
     @PreAuthorize("hasAuthority('ADMIN_USER')")
     @RequestMapping(value = "/private/championnat", method = RequestMethod.DELETE)
     public void deleteChampionnat(@RequestParam Long id){
-        Championnat championnat = new Championnat();
-        championnat.setId(id);
-        //TODO : idealement transactionnel...
-        divisionRepository.deleteByChampionnat(championnat);
-        championnatRepository.delete(id);
+        if (isCalendrierDeletable(id)){
+            Championnat championnat = new Championnat();
+            championnat.setId(id);
+            divisionRepository.deleteByChampionnat(championnat);
+            championnatRepository.delete(id);
+        }
     }
 
     @PreAuthorize("hasAuthority('ADMIN_USER')")
